Reorganize binary structs
